GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - factorio-latest (refs/gitbutler/factorio-latest)
   - flake.lock
   - flake.nix
   - modules/nixos/archetypes/gaming/default.nix
 - hyprpanel (refs/gitbutler/hyprpanel)
   branch head: acde84ab10891f923177d3dde85b826296a1b4bc
   - modules/nixos/desktop/addons/hyprpanel/options.json
   - modules/nixos/desktop/addons/hyprpanel/default.nix
   - modules/nixos/desktop/addons/hypridle/default.nix
 - fix hyprland nvidia issue (refs/gitbutler/fix-hyprland-nvidia-issue)
   branch head: 5158f2a21e773e01f3d4f296db87517bf167c7b7
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch